requirements/dev.txt
django-extensions==1.9.6
ipython
django-debug-toolbar==1.8
